Bugfixes. A lot of them. Some refactoring. Still under way.

+++ lib/action_routes.cc
@@ -140,7 +140,7 @@ namespace xameleon {
         segment.clear();
         url.rest = url.path + i;
 
-        int sz = chunks.size();
+        int sz = (int) chunks.size();
         for (i = 0; i < sz; i++) {
             segment = chunks[i];
             if (node->_child_points.count(segment) == 0) {
